requirements2.txt
# Core app framework
fastapi==0.115.6          # API layer
uvicorn==0.35.0           # ASGI server
streamlit==1.38.0         # Web UI

# LLM and Vector DB
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.28
langgraph==0.6.4
chromadb==1.0.16
ollama==0.5.3
openai==1.99.3

# Data processing
pandas==2.3.1
numpy==2.3.2

# Database
SQLAlchemy==2.0.42
psycopg2-binary==2.9.9     # Postgres driver
sqlite-utils==3.36         # Optional for SQLite ops

# Utils
python-dotenv==1.1.1
requests==2.32.4
tqdm==4.67.1